requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcc==0.18.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
contourpy==1.3.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.15
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.129.0
fastjsonschema==2.21.1
filelock==3.24.2
flatbuffers==25.12.19
fonttools==4.59.0
fqdn==1.5.1
fsspec==2026.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==1.1.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hidpidaemon==18.4.6
hpack==4.1.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.30.0
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kaggle==1.7.4.5
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
lark==1.2.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
loguru==0.7.3
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mmh3==5.2.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.4.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.0
onnxruntime==1.23.2
openai==2.21.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pop-transition==1.1.2
portalocker==3.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.33.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.5
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydbus==0.6.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-gnupg==0.4.8
python-json-logger==3.3.0
python-slugify==8.0.4
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==27.0.0
google-genai==1.15.0
qdrant-client==1.16.2
referencing==0.36.2
repolib==2.2.1
repoman==1.4.0
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rank-bm25==0.2.2
rich==14.3.2
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.15.3
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
Send2Trash==1.8.3
sessioninstaller==0.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
systemd-python==234
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.23.1
typer-slim==0.23.1
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uri-template==1.3.0
urllib3==2.6.3
uvicorn==0.40.0
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
xdg==5
xkit==0.0.0
zipp==1.0.0
